backend/pyproject.toml
[tool.poetry]
name = "plagiarism-detection-backend"
version = "0.1.0"
description = "Production-grade backend for the plagiarism detection system."
authors = ["Jules"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"

# Core FastAPI and Web Server
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
gunicorn = "^21.2.0"
pydantic-settings = "^2.1.0"

# Database and ORM
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.12.1"
asyncpg = "^0.29.0"
pgvector = {extras = ["sqlalchemy"], version = "^0.2.0"}
"psycopg2-binary" = "^2.9.9"

# Asynchronous Tasks
celery = "^5.3.6"
redis = "^5.0.1"

# Authentication and Security
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
boto3 = "^1.28.57"

# File Handling and Parsing
python-multipart = ">=0.0.7"
"pdfminer.six" = "^20221105"
"python-docx" = "^1.1.0"
minio = "^7.1.14"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"

# Reporting and Monitoring
reportlab = "^4.0.7"
loguru = "^0.7.2"
prometheus-fastapi-instrumentator = "^6.0.0"

[tool.poetry.group.ai.dependencies]
# AI/ML Models and Transformers
sentence-transformers = "^2.2.2"
open_clip_torch = "^2.20.0"
transformers = "^4.35.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
ruff = "^0.1.6"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
